hw/vfio/region: Create dmabuf for PCI BAR per region
Linux kernel now has a dmabuf exporter to expose PCI device memory for P2P
use case. Set it up properly per region once the mmap is done. Feed the fd
to ram_block->fd so as to go via dma_map_file for iommufd in the kernel to
import the dmabuf to map the BAR memory properly in the page table.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>